+++ src/__tests__/withStyles.types-test.tsx
@@ -1,4 +1,4 @@
-import React, { FC, JSX, forwardRef } from "react";
+import React, { JSX, FC, forwardRef } from "react";
 import { TextInput } from "react-native";
 
 import { makeUseStyles } from "../makeUseStyles";