+++ src/withStyles.tsx
@@ -1,4 +1,11 @@
-import type { ComponentProps, ComponentRef, ComponentType, Ref } from "react";
+import type {
+  ComponentProps,
+  ComponentRef,
+  ComponentType,
+  ForwardRefExoticComponent,
+  PropsWithoutRef,
+  RefAttributes,
+} from "react";
 import { forwardRef } from "react";
 
 interface InjectedStyledProps<Styles> {
@@ -11,20 +18,16 @@ export type InferInjectedStyledProps<Fn extends (...args: any) => any> =
 export const withStyles = <Styles,>(useStyles: () => Styles) => {
   const WithStyles = <TComponent extends ComponentType<any>>(
     Component: TComponent,
-  ) => {
-    const renderComponent = (
-      props: Omit<ComponentProps<TComponent>, "styles">,
-      ref: Ref<ComponentRef<TComponent>>,
-    ) => {
+  ): ForwardRefExoticComponent<
+    PropsWithoutRef<Omit<ComponentProps<TComponent>, "styles">> &
+      RefAttributes<ComponentRef<TComponent>>
+  > => {
+    const WithStylesComponent = (props: any, ref: any) => {
       "use memo";
-      // eslint-disable-next-line react-hooks/rules-of-hooks
       const styles = useStyles();
-      // @ts-expect-error: 'ref' as never
       return <Component {...props} ref={ref} styles={styles} />;
     };
-
-    // @ts-expect-error
-    return forwardRef(renderComponent);
+    return forwardRef(WithStylesComponent);
   };
 
   return WithStyles;